Reworked to use engine param2 colorization
Uses the place-then-paint method to make colors. To get a regular color,
place some Moreblocks super glow glass, then right click on it with some
dye to color it.  To get the pastel colors, right-click the colored
glass with Moreblocks' "sweeper" to "brush off" some of the color.  Do
it again to change pastel to faint.

Right click a pastel or faint with some dye to re-color it.  You have to
dig and re-place if you want to go darker/more saturated again.